Fixed UTF8 bug and have loading spinner on game
Before sometimes my get_phrase.php script would return a empty string
as it's json. This was because not all of my strings in my phrases are
utf8 encoded. I changed that now with charset=utf8 on my PDO. I also now
have a spinner for when I am loading a phrase from the backend