Fixed bugs with the leaderboard
A person could previously go to the leaderboard without a phrase, and not be given an error. Now they get a helpful error message explaining to them that they need to play a game. Previously there was still an insertion to the database if the current score was less than the previous. Now I prevent that. The board part of leaderboard is done.