Built the skeleton for the admin feature
I added a boolean admin field into Account table, and added a corresponding attribute to the Account class. I created the various scripts for the features an admin can do: create a new phrase, make an additional admin account, and configure the backend. There is button on main.html that allows only admin accounts to access the admin page.